Box2d Support Shape Box #138

+++ engine/core/util/base64.cpp
@@ -68,7 +68,7 @@ namespace Echo
 		if (data[input_length - 2] == '=') output_length--;
 
 		m_decoded.resize(output_length);
-		for (int i = 0, j = 0; i<input_length;)
+		for (size_t i = 0, j = 0; i<input_length;)
 		{
 			uint32_t sextet_a = data[i] == '=' ? 0 & i++ : decoding_table[data[i++]];
 			uint32_t sextet_b = data[i] == '=' ? 0 & i++ : decoding_table[data[i++]];
